engine/requirements.txt
fastapi
uvicorn
python-socketio
playwright
dashscope
python-dotenv
